# Add Encode(Span<char>) API