… src/System.Text.Encodings.Web/tests/JavaScriptStringEncoderTests.cs
… src/System.Text.Json/tests/Utf8JsonWriterTests.cs