@steveharter 2019-08-05
> cc @ericstj
> FWIW the ASCII escaping in System.Text.Json is ~25% faster than System.Text.Encoding.Web.JavaScriptEncoder - thus the reason why there is some duplicate functionality between the two; S.T.J will only call S.T.E.W for non-ASCII or when using a non-default encoder.
>> What's the reason for that? Is there a way for us to bring some of this perf improvement into S.T.E.W itself?

The reason S.T.E.W is slower is because there is a class hierarchy (JavaScriptEncoder : TextEncoder) where both the default encoder and custom share the same (slower) code which is driven off of a combination of valid and invalid ranges which are flexible and can be specified by the caller -- and this code can't have the micro-optimizations that we have in S.T.J.

However we could create a new internal type (DefaultJavaScriptEncoder in S.T.E.W) and move code from S.T.J to S.T.E.W however using a custom encoder is still going to be 25% slower since it won't use that fast code.

Also note that S.T.E.W was already made 3x-4x faster from the original (for Utf8).