@GrabYourPitchforks 2019-08-05
It'd be good to have a unit test validating the changes for the latest iteration https://github.com/dotnet/corefx/pull/39900/commits/390e6c77528db10aef5dce208dcac2196903ef8e, but otherwise LGTM. Thanks!